+++ window/src/os/wayland/mod.rs
@@ -11,6 +11,7 @@ mod copy_and_paste;
 mod drag_and_drop;
 // mod frame;
 mod data_device;
+mod fractional_scale;
 mod keyboard;
 mod pointer;
 mod seat;